[FIX] website_sale: guard against empty product in GA4 tracking data
**Steps to reproduce**
- Configure a Google Analytics key on the website.
- Type in the search bar on /shop (autocomplete dropdown).
- A `ValueError: Expected singleton: product.product()` is raised.

The same crash occurs when loading a dynamic product snippet in
"hide variants" mode, or when opening a product page for a template
whose selected combination has no materialized product.product variant.

**Cause**
`_get_combination_info` intentionally sets `product` to an empty
`product.product` recordset when `only_template=True` is passed, or
when `_get_variant_for_combination` finds no realized variant for the
given combination. The GA branch then calls `_get_google_analytics_data`
with that empty recordset, which will cause a crash because
`product._split_standard_from_custom_attributes()` expects a singleton.

**Change**
Move `_split_standard_from_custom_attributes` from `product.product`
to `product.template.attribute.value` so it operates on a recordset
without requiring a singleton. Call it directly on
`combination_info["combination"]` in `_get_google_analytics_data`,
which is always a PTAV recordset and is empty when no variant is
resolved.

opw-6222822
