+++ addons/website_sale/models/product_template_attribute_value.py
@@ -2,6 +2,8 @@
 
 from odoo import models
 
+from odoo.addons.website_sale import const
+
 
 class ProductTemplateAttributeValue(models.Model):
     _inherit = "product.template.attribute.value"
@@ -23,3 +25,26 @@ class ProductTemplateAttributeValue(models.Model):
             price_extra = self.currency_id._convert(from_amount=price_extra, to_currency=currency)
 
         return self.product_tmpl_id._apply_taxes_to_price(price_extra, currency)
+
+    def _split_standard_from_custom_attributes(self):
+        """Split PTAVs into directly mapped fields and the rest.
+
+        Direct fields are attributes whose external_identifier matches a known
+        identifier used across Microdata, GMC Feeds, and Tracking.
+
+        :return: A tuple of:
+            - dict of {external_identifier: value} for known direct fields
+            - dict of {attribute_name: value} for all other attributes
+        :rtype: tuple(dict, dict)
+        """
+        direct = {}
+        others = {}
+        for ptav in self:
+            external_id = ptav.attribute_id.external_identifier
+            ext_id = external_id and external_id.lower()
+            value = ptav.product_attribute_value_id.name
+            if ext_id and ext_id in const.DIRECT_MAPPED_ATTRIBUTE_IDENTIFIERS:
+                direct[ext_id] = value
+            elif external_id:
+                others[external_id] = value
+        return direct, others
